se movió la función verifyToken a auth.controller.js y el endpoint verifytoken a auth.routes.js
