+++ controllers/user.controller.js
@@ -61,15 +61,4 @@ const create = async (req, res) => {
     }
   };
 
-  const verifyToken = async (req, res) => {
-    try {
-      // retorna los datos excluyendo el password
-      usuarioVerificado = await User.findById(req.user.id).select('-password');
-      res.json(usuarioVerificado);
-    } catch (error) {
-      console.error(error);
-      res.status(500).json({ message: 'error verificando el token'});
-    }
-  };
-
-module.exports = { readAll, readOne, create, update, remove, verifyToken };
+module.exports = { readAll, readOne, create, update, remove };
